# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Luna AI API",
    "version": "1.0.0",
    "description": "API Documentation for Luna AI Streamer - AI-powered streamer with VTube Studio integration, emotion system, and Rock Paper Scissors game",
    "contact": {
      "name": "Luna AI Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/luna/rps/queue": {
      "post": {
        "tags": ["RPS"],
        "summary": "Queue for PvP match",
        "description": "Join the matchmaking queue for PvP game",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet"],
                "properties": {
                  "wallet": {
                    "type": "string",
                    "description": "Wallet address"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Queued successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/rps/match": {
      "get": {
        "tags": ["RPS"],
        "summary": "Get match status",
        "description": "Check if a match has been found",
        "parameters": [
          {
            "name": "wallet",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "description": "Match status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "matched": {
                      "type": "boolean"
                    },
                    "matchId": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/luna/rps/submit": {
      "post": {
        "tags": ["RPS"],
        "summary": "Submit RPS choice",
        "description": "Submit Rock Paper Scissors choice for a match",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["wallet", "matchId", "choice"],
                "properties": {
                  "wallet": {
                    "type": "string"
                  },
                  "matchId": {
                    "type": "string"
                  },
                  "choice": {
                    "type": "string",
                    "enum": ["rock", "paper", "scissors"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Choice submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "submitted": {
                      "type": "boolean"
                    },
                    "bothSubmitted": {
                      "type": "boolean"
                    },
                    "result": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "AdminSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-secret",
        "description": "Admin secret key (set ADMIN_SECRET in .env)"
      },
      "PurchaseSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-purchase-secret",
        "description": "Purchase webhook secret (set PURCHASE_SECRET in .env)"
      }
    }
  }
}